[NGRINDER-218] clone test error during running
When the test is running, it will get $("#testId").val()  and update the
status. But if we clone the test, the $("#testId") will be set as empty.
So I saved the $("#testId").val()  as global var to avoid this kind of
problem.

And and modified some places to avoid looking for jquery element every
time.