[NGRINDER-269] error for monitoring
If there are 2 tests, aiming at same target server, there will be 2 set
of records in DB. Then it will cause the monitor data improper.
I monified as below:
Use a manager to start monitor worker for every agents, if there are 2
tests monitoring one target, there is a reference counter for that.